fix(mobile): change UserResponse heightCm/weightKg from Float to Double
Kotlin/Native (iOS) throws IllegalArgumentException when deserializing
JSON numbers (always Double) into Float fields. Also add
@OptIn(ExperimentalCoroutinesApi::class) to ViewModel tests.

+++ mobile/shared/src/commonMain/kotlin/ru/skatelab/shared/api/UsersApi.kt
@@ -14,7 +14,7 @@ class UsersApi(private val client: HttpClient) {
         displayName: String? = null,
         bio: String? = null,
         heightCm: Int? = null,
-        weightKg: Float? = null,
+        weightKg: Double? = null,
     ): UserResponse =
         client.patch("/users/me") {
             contentType(ContentType.Application.Json)
